Reorganize platform specific source files (#36)
* Reorganize kotlin package structure

* Reorganize swift source structure

* Create directory for custom types

For types extending native counterparts from AppleMaps
