+++ expo-maps/android/src/main/java/expo/modules/maps/interfaces/Markers.kt
@@ -1,4 +1,6 @@
-package expo.modules.maps
+package expo.modules.maps.interfaces
+
+import expo.modules.maps.MarkerObject
 
 interface Markers {
   fun setMarkers(markerObjects: Array<MarkerObject>)
